[CI TEST, DO NOT MERGE] extract stack exhaustion test
